Fix duplicate target bug
It was pretty easy to accidentally add the same target twice to the
list so I added a safety net to the makefile generator.  Note that we
still need to prevent circular dependencies from getting into the main
dependency stack, and it's best to skip dependencies whenever possible,
so this might not actually be the best idea.