Add "pandoc" support to tek
This generates a HTML file for every LaTeX file.  The current support
isn't all that good, but it does at least generate decent HTML files
for some small subset of LaTeX so I guess it's better than nothing.  I
would be wonderful if it was possible to hack LaTeX to make pandoc
work better, I'll have to look into that...