Allow inkscape to convert SVG files
It seems that inkscape generates a background inside its SVG files,
which is a bit of a pain -- it means that pdfcrop won't remove the
white borders from them.  This produces ugly looking PDF files.

The remedy seems to be to use inkscape to natively generate PDF files,
which I guess will probably also mean better support.  inkscape has a
commandline option to not generate the border when exporting, which
allows compact PDF files to be generated.

In order to force inkscape to generate PDF files instead of
imagemagick, name your file FILENAME.inkscape.svg.