Redirect gnuplot's stderr to /dev/null
This may or may not be a good idea: essentially what happens is that
when you try to fit something to a plot, gnuplot will output a whole
bunch of junk to stderr.  As far as I can tell there's no way to get
around this, which is a huge pain.

This unfortunately means that you won't see gnuplot error messages any
more.