Add support for LaTeX's index commands
This searches through every input .tex file looking for a
"\printindex" statement.  If one is found, a run of makeindex is added
to the Makefile rule for that target (along with an extra pdflatex
run, for good measure).  This should allow indicies to actually be
generated into the output PDF.