Be atomic about generating new PDFs
I used to use "cp" to speed up the process of overwriting
"pdflatex"-generated PDF files.  Essentially the problem is that
there's a race condition between the PDF reader getting the notify
call and pdflatex finishing.  Since cp is so much faster than pdflatex
this hid the race condition, but the better way is to just use
rename() to make this actually be atomic.  On my machine "mv" uses
rename(), so this appears to be fine.

This should solve any reader update problems on slow systems.